# Speed boost for DO in UI via Camshaft OBS_GetMeta/OBS_GetData