--- a/lib/node/nodes/data-observatory-measure.js
+++ b/lib/node/nodes/data-observatory-measure.js
@@ -25,12 +25,32 @@ DataObservatoryMeasure.prototype.sql = function() {
 };
 
 var queryTemplate = Node.template([
-    'SELECT',
-    '  {{=it.columns}},',
-    '  cdb_dataservices_client.OBS_GetMeasure(' +
-    '    the_geom,' +
-    '    \'{{=it.segment_name}}\'{{?it.percent}},',
-    '    \'{{=it.percent}}\'{{?}}',
-    '  ) AS {{=it.final_column}}',
-    'FROM ({{=it.source}}) AS _camshaft_do_measure_analysis'
+    'WITH summary AS ( ',
+    '  SELECT ST_SetSRID(ST_Extent(the_geom), 4326) extent, ',
+    '         count(*)::INT numgeoms, ',
+    '         ARRAY_AGG(DISTINCT ST_GeometryType(the_geom)) AS geom_types, ',
+    '         ARRAY_AGG((the_geom, cartodb_id)::geomval) geomvals ',
+    '  FROM {{=it.source}} ',
+    '  ), ',
+    'meta AS ( ',
+    '  SELECT cdb_dataservices_client.OBS_GetMeta( ',
+    '            extent, ',
+    '            (\'[{"numer_id": "{{=it.segment_name}}"}]\')::JSON ',
+    '          , null, null, numgeoms) as meta FROM summary ',
+    '  ), ',
+    'data AS ( ',
+    '  SELECT id, (data->0->>\'value\')::Numeric {{=it.final_column}} ',
+    '  FROM cdb_dataservices_client.OBS_GetData( ',
+    '    (SELECT geomvals FROM summary), ',
+    '    (SELECT jsonb_set(meta::jsonb, \'{0,normalization}\', ',
+    '        CASE ',
+    '          WHEN meta->0->>\'numer_aggregate\' = \'sum\' AND geom_types = \'{ST_Point}\' ',
+    '            THEN \'"area"\'::jsonb ',
+    '          ELSE \'"prenormalized"\'::jsonb ',
+    '        END ',
+    '    )::json FROM meta, summary) ',
+    '  )) ',
+    'SELECT {{=it.columns}}, data.{{=it.final_column}} ',
+    'FROM data, {{=it.source}} source ',
+    'WHERE source.cartodb_id = data.id '
 ].join('\n'));