--- a/test/fixtures/cdb_dataservices_client/obs_getmeasure.sql
+++ b/test/fixtures/cdb_dataservices_client/obs_getmeasure.sql
@@ -1,26 +1,3 @@
-CREATE OR REPLACE FUNCTION cdb_dataservices_client.OBS_GetMeasure(
-  geom geometry,
-  segment_name text
-)
-RETURNS numeric
-AS $$
-BEGIN
-  RETURN trunc(random() * 1e5 + 1);
-END;
-$$  LANGUAGE plpgsql;
-
-CREATE OR REPLACE FUNCTION cdb_dataservices_client.OBS_GetMeasure(
-  geom geometry,
-  segment_name text,
-  denominator text
-)
-RETURNS numeric
-AS $$
-BEGIN
-  RETURN trunc(random() * 100 + 1);
-END;
-$$  LANGUAGE plpgsql;
-
 CREATE OR REPLACE FUNCTION cdb_dataservices_client.OBS_GetMeta(
   geom_ref Geometry(Geometry, 4326),
   params json,
@@ -35,7 +12,7 @@ BEGIN
 END;
 $$ LANGUAGE plpgsql;
 
-CREATE OR REPLACE FUNCTION cdb_dataservices_client.obs_getdata(
+CREATE OR REPLACE FUNCTION cdb_dataservices_client.OBS_GetData(
   geomvals geomval[],
   params json,
   merge boolean DEFAULT true