## Summary

Update of `data-observatory-measure.js` to take advantage of `obs_getdata/obs_getmeta` to have some serious performance improvements.  This is a drop-in replacement that does not require any UI changes.

## Why we are doing this

- Performance in the existing analysis interface is very slow, on the order of 20-30QPS.
- With the deployment of several new dataservices/observatory functions (see requirements) we can use much faster methods for obtaining Observatory data

## Requirements

- https://github.com/CartoDB/dataservices-api/pull/339 must be merged and confirmed to function as expected in user accounts

## Statistics

test case | old time | new time | old qps | new qps | speedup
--- | --- | --- | --- | --- | ---
get population for 1000 US census tract polygons | 48.6s | 1.7s | 20.5 | _588_ | *28x*
get median income for 1000 US points | 47.2s | 0.2s | 21.1 | _5000_ | *236x*
get population density for 1000 US points | 47.2s | 0.2s | 21.1 | _5000_ | *236x*