@talos 2017-01-31
@saleiva This results in some very serious performance improvements, do you think you can fit this into the roadmap sometimes soon?

Tests are failing right now, but I suspect that is at least in part because obs_getdata and obs_getmeta are not part of the dataservices-api on the test box.