@javisantana 2017-02-01
- Could you explain what obs_getdata/obs_getmeta is? I couldn't find any documentation. Is this some kind of batch thing?
- The first CTE could be killer, aggregate all the geometries for a table in an array is kind of hard. Did you do some test over tables with big polygons?