@talos 2017-02-01
* Documentation is in progress, but you can find a sample here:
  - https://github.com/CartoDB/observatory-extension/blob/obs-getdata-getmeta-docs/doc/measures_functions.md#obs_getmetaextent-geometry-metadata-json
  - https://github.com/CartoDB/observatory-extension/blob/obs-getdata-getmeta-docs/doc/measures_functions.md#obs_getdatageomvals-arraygeomval-metadata-json
  - https://github.com/CartoDB/observatory-extension/blob/obs-getdata-getmeta-docs/doc/measures_functions.md#obs_getdataids-arraytext-metadata-json

* The CTE will pose problems with lots of large geometries, but also having run some tests on large geometries prior to this CTE I'm pretty sure we run into bottlenecks at other points calculations interpolations for such complex geometries before we hit bottlenecks putting everything into the array.

I was able to send 1M points through the PL/Proxy hoops, but I need to do tests to see at what point the array becomes too large with polygons.

In the interest of being able to deploy anything, I would be in favor of moving forward provide we don't actively degrade performance from what we have now.  If you look at `OBS_GetMeasure` for complex geometries (several thousand vertices per polygon) as hit in the [OBS as of 1.2.1](http://52.71.151.140/perftest/#00825e4ba115bc9d435a138ab19983c211d24f6e), we can handle ~1 per second.

<img width="889" alt="screen shot 2017-02-01 at 10 52 06 am" src="https://cloud.githubusercontent.com/assets/415831/22514186/7f284418-e86c-11e6-882b-e60436a275d7.png">

If we can fit 100 such complex geometries into an array without hitting any errors, this is still essentially an improvement from current conditions.

The longer term solution would be to rewrite the camshaft node to chunk the requests into roughly equal-point calls to `OBS_GetData`, which would get us the best of both worlds.

I'll do some tests on how we perform with arrays of large geometries today and update.