@rochoa 2017-02-01
> The longer term solution would be to rewrite the camshaft node to chunk the requests into roughly equal-point calls to OBS_GetData, which would get us the best of both worlds.

That assumes camshaft allows to chunk data. But I don't see that happening with the current implementation/limitations camshaft has: https://github.com/CartoDB/camshaft/blob/44bfe26100a49c394c8cca2fc51e42050ffdd15a/docs/INTERNALS.md#internals.