@talos 2017-02-01
I'm using a liberal interpretation of the word "chunk".

I could see some fancy SQL that can do the `array_agg` of a user's rows with a sensitivity to the size of the arrays that are being aggregated.  `OBS_GetData` would then be called on each row separately, avoiding overloading the length of the array at any point.  It would all still be contained in one query, and technically all happening at once.