@talos 2017-02-01
@rafatower @javisantana Performance tests for large geomval arrays run through `cdb_dataservices_client` to the observatory can be found [here](https://gist.github.com/talos/0e4d49619cb3a3778999dc1d5f888018).

The tests should be replicated on a production or staging server so that we fully mimic the way data is ferried.  OTOH, these tests did use PL/Proxy and are promising that we can work with several thousand large and several dozen huge geometries without having issues merely for constructing the array.  The emergent issues are down the line when we actually do interpolations and intersections, and were much worse with the old setup.