@javisantana 2017-02-02
Thanks @talos, really useful

One interesting topic (maybe not for this time) would to to try to use TWKB. Instead of sending an array of (id, geometry) use st_astkwb, from postgis doc:

```sql
SELECT ST_AsTWKB(array_agg(geom), array_agg(gid)) FROM mytable;
                 st_astwkb
--------------------------------------------
\x040402020400000202
```

TKWB stores id inside the geometry, it uses 4x less memory and BW and is standard. The only downside is the loss of precision in some cases but not sure if the results are going to be much different. Another good thing is with tables with lot of points, doing:

```sql
SELECT ST_AsTWKB(array_agg(the_geom), array_agg(cartodb_id)) FROM (select cartodb_id, the_geom from table order by the_geom) alias
```

Will improve the payload size for those tables since the delta encoding will do its work. 

It'd improve much more if the transfer would compress the data but not sure pl/proxy is doing that (cc @rafatower )