@rafatower 2017-02-02
@talos thanks for the report. A couple questions: what kind of failures did you get with obs_getmeta/obs_getdata? what results (or absence of results) did you get in the blank cells?