@rochoa 2017-02-02
> One interesting topic (maybe not for this time) would to to try to use TWKB. Instead of sending an array of (id, geometry) use st_astkwb, from postgis doc:

That's interesting for some other cases where we use a similar approach, kmeans for instance.