@talos 2017-02-02
I actually think this is an ideal use case for TWKB -- clearly this is what it was meant for.

I'll do a few tests.  I think if it doesn't cause major validity issues (a possible consequence of naive loss of precision, although conversion to TWKB and precision loss would be up to the client) it would be worse using TWKB instead of geomval from the get-go.

@rafatower the failures were unfortunately masked by the cdb_dataservices_client wrapper.  I didn't have time to re-run those on cdb_observatory directly, which would have revealed the error.  On a separate note, is there any reason why we shouldn't try to expose the internal error at least on the log, if not to the user?

The blank calls I just didn't run because the times were clearly expanding in such a manner they would take too long for me to put the document together before EOD and include them.