@rafatower 2017-02-02
> @rafatower the failures were unfortunately masked by the cdb_dataservices_client wrapper. I didn't have time to re-run those on cdb_observatory directly, which would have revealed the error.

@talos when did you execute those tests? what user did you use? we have rollbar and the logs and we shall check them.

> On a separate note, is there any reason why we shouldn't try to expose the internal error at least on the log, if not to the user?

Security reasons: do not leak sensitive info (e.g: 3rd party api keys in a stack trace, for instance).

Have you run any specific tests to mitigate the risk of invalid geoms when simplifying?