@talos 2017-02-02
@rafatower I have a local installation of dataservices-api I use for development, so the logs would not appear on rollbar.

We have unit tests with complicated geometries to make sure we don't generate anything invalid.  However, it can still happen as it's tricky to mitigate.  It's possible that's what happened.