@rafatower 2017-02-02
In a local setup you can do a number of things:
- tweak postgres conf, reload/restart and check the logs dumped in `/var/log/postgresql`
- configure the logger in your dataservices-api server:
https://github.com/CartoDB/dataservices-api
```
SELECT CDB_Conf_SetConf('logger_conf', '{"geocoder_log_path": "/tmp/geocodings.log", [ "min_log_level": "[debug|info|warning|error]", "rollbar_api_key": "SERVER_SIDE_API_KEY", "log_file_path": "LOG_FILE_PATH"]}');
```
set `LOG_FILE_PATH` to something convenient to you and watch there.